test(toggle): migrate item
