+++ core/src/components/toggle/test/item/label.e2e.ts-snapshots/toggle-inset-list-ios-ltr-Mobile-Chrome-linux.png
