+++ core/src/components/toggle/test/item/label.e2e.ts-snapshots/toggle-inset-list-ios-rtl-Mobile-Safari-linux.png
