+++ core/src/components/toggle/test/item/label.e2e.ts-snapshots/toggle-inset-list-md-ltr-Mobile-Safari-linux.png
