+++ core/src/components/toggle/test/item/label.e2e.ts-snapshots/toggle-item-color-ios-ltr-Mobile-Firefox-linux.png
