+++ core/src/components/toggle/test/item/label.e2e.ts-snapshots/toggle-item-color-md-ltr-Mobile-Safari-linux.png
