+++ core/src/components/toggle/test/item/label.e2e.ts-snapshots/toggle-list-md-ltr-Mobile-Firefox-linux.png
