+++ core/src/components/toggle/test/item/label.e2e.ts-snapshots/toggle-list-md-rtl-Mobile-Chrome-linux.png
